Refactor contacts permission handling
